backend/requirements/base.txt
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.7.0
django-environ==0.4.5
dj-database-url==0.5.0
psycopg2==2.8.6
whitenoise==5.2.0
